separated helpers for tests from test functions
