+++ tests/quantization/test_precision_init.py
@@ -47,7 +47,7 @@ from tests.quantization.test_quantization_helpers import compare_multi_gpu_dump,
     get_quantization_config_without_range_init, distributed_init_test_default, post_compression_test_distr_init, \
     get_squeezenet_quantization_config
 from tests.test_compressed_graph import check_graph
-from tests.test_helpers import create_compressed_model_and_algo_for_test, MockModel, create_conv, \
+from tests.helpers import create_compressed_model_and_algo_for_test, MockModel, create_conv, \
     create_mock_dataloader
 
 
@@ -411,6 +411,7 @@ def test_hawq_manual_configs(manual_config_params, hw_config):
     if hw_config:
         config['hw_config'] = hw_config.value
     model = load_model(config['model'], pretrained=False)
+    model.eval()
 
     _, compression_ctrl = create_compressed_model_and_algo_for_test(model, config)
