+++ tests/sparsity/magnitude/test_modules.py
@@ -19,7 +19,7 @@ from nncf.layers import NNCFConv2d, NNCFLinear
 from nncf.module_operations import UpdateWeight
 from nncf.sparsity.layers import BinaryMask
 from nncf.sparsity.magnitude.functions import normed_magnitude, abs_magnitude, calc_magnitude_binary_mask
-from tests.test_helpers import fill_conv_weight, fill_linear_weight, fill_bias
+from tests.helpers import fill_conv_weight, fill_linear_weight, fill_bias
 
 
 class TestModel(nn.Module):
