+++ tests/test_compressed_graph.py
@@ -33,7 +33,7 @@ from nncf.utils import get_all_modules_by_type
 from tests import test_models
 from tests.modules.seq2seq.gnmt import GNMT
 from tests.modules.test_rnn import replace_lstm
-from tests.test_helpers import get_empty_config, create_compressed_model_and_algo_for_test
+from tests.helpers import get_empty_config, create_compressed_model_and_algo_for_test
 from nncf.quantization.algo import PotentialQuantizedModule
 from nncf.quantization.quantizer_propagation import QuantizerPropagationSolver
 from nncf.quantization.layers import QuantizerConfig
